docker-airflow-run/requirements.txt
jupyterlab
deepchecks
pandas
numpy
matplotlib
flask
scikit-learn
scipy
xgboost
statistics
sqlalchemy
sqlalchemy_utils
psycopg2-binary
apache-airflow